break: move `reply` to router module
