+++ src/response.d.ts
@@ -1,9 +1,5 @@
 /// <reference lib="webworker" />
 
-export type FetchHandler = (event: FetchEvent) => void;
-export type ResponseHandler = (event: FetchEvent) => Promise<Response> | Response;
-export function reply(handler: ResponseHandler): FetchHandler;
-
 export type HeadersObject = Record<string, string>;
 
 export declare class ServerResponse {
