+++ types/check.ts
@@ -1,13 +1,12 @@
 import * as Cache from 'worktop/cache';
 import { Database, until } from 'worktop/kv';
-import { Router, STATUS_CODES } from 'worktop';
-import { reply, ServerResponse } from 'worktop/response';
+import { ServerResponse } from 'worktop/response';
+import { reply, Router, STATUS_CODES } from 'worktop';
 import { byteLength, HEX, uid, uuid } from 'worktop/utils';
 
 import type { KV } from 'worktop/kv';
+import type { FetchHandler, Route } from 'worktop';
 import type { ServerRequest, IncomingCloudflareProperties } from 'worktop/request';
-import type { FetchHandler } from 'worktop/response';
-import type { Route } from 'worktop';
 
 declare function assert<T>(value: T): void;
